+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 0.4.0 - 2026-09-12
+
 ### Changed
 
 - Change default value for bundle configuration option `pretty_print` to `'%kernel.debug%'`.